use render component: syntax
In https://github.com/github/actionview-component/issues/21,
@nogtini pointed out that our existing syntax of calling
`render(MyComponent.new)` was incompatible with the
`render(... collection: @products)` syntax Rails allows for partials.

In order to support this, and potentially other changes to align
with the existing Rails API, it is necessary to change our
syntax used to render components in views.

Now, components are rendered as:
`render(component: MyComponent, locals: { foo: :bar })`.

As this is a major breaking change, I feel this change should bump
us straight to v2.0.0.