In #21, @nogtini pointed out that our existing syntax of calling `render(MyComponent.new)` makes it difficult to support rendering collections in the way that Rails allows one to do so with partials.
In order to support rendering collections using components,
and potentially other changes to align with the existing Rails API,
it is necessary to change our syntax used to render components in views.

Now, components are rendered as:
`render(MyComponent, locals: { foo: :bar })`.

The old syntax of:
`render(MyComponent.new(foo: :bar)` is still supported, but will be
deprecated in v2.