Handle components with partials that are rendered elsewhere
Our application renders a particular partial in a
component and also in regular views, and this was causing a method
missing error on Rails 6 as the component didn't have access to the
compiled partial method.

To resolve it we had to store the view context, override the
compiled_method_container to use the same one as the view context,
and delegate any missing methods (e.g. compiled partials) to the view
context.