explicitly remove extension from path vs splitting on '.'
Our existing implementation of looking up sibling template files
split the source_location path on '.', assuming that the .rb
file extension was the only place a period was used in the entire
file path. This meant that any component file in a path that contained
other periods failed to find its sibling templates.

I did not include a test here as it seemed difficult to impossible to write one.

[Fixes: #83]

Co-Authored-By: Anton <profox.rus@gmail.com>