Removing "components" from virtual path
Rails default for translation shortcut paths do not include the parent folder where the translations are used.
For instance, when you use `t('.success')` in a controller, the translation key does not include the `controllers` folder.
Therefore, the word `components` get removed from the virtual path of components.
This will allow us to use `t('.title')` in components and expect the translation key not to include the `components` key as the top level key in the translation yml file.

What happens when the developer uses a different folder for components? The virtual path will just be the location of the file, in other words, the output of `Class#source_location`.