Avoid dropping elements in the render_inline test helper and adjust tests
The main thrust of this change is to avoid dropping elements by using
`.css("body > *")`, which will strip out some elements in some
situations and produce unexpected test results. At the same time, we'll
clean up some tests to ensure that whitespace stripping is more
consistent and test arguments are given in a more consistent order.