Add from_options_with_default to Base
from_options_with_default is a lightweight helper that
enables components to gracefully fall back to a default
value when invalid input is provided.

We're starting to move away from using validations in our
components internally, as we've found that what we really
need is a guarantee that a value will be one that we expect.
What we don't need is for our view code to raise exceptions
when an unexpected value is passed.

In our ~50 components, we've only used two validation methods:
presence and inclusion. We use this helper in place of
inclusion validations.