Set template filepath and line number for better backtraces
Currently `__FILE__` and `__LINE__` are the arguments passed to
`class_eval` which means when an exception is raised in the template it
will point back to the component instead of the correct line of the
template.

This updates the filepath and line number arguments passed to
`class_eval` to point to the template instead of the component.

Since Rails uses the first template found in the backtrace as the
displayed source this will now show where the exception actually
occurred instead of the call that rendered the component.