Allow to access params in preview examples
This is helpful when you want to show dynamic values on the component
preview.

One use case is integrating Storybook and the Knobs addon, which allows
to tweak the values directly from the component's page.

We can take those values and pass them down as params to the Rails
preview page, which will then be used to set the values on the
component.