Fix block based form helper rendering
In Rails, when you create a form it saves a reference to the current
template object that is rendering the form. When you render a form
builder helper method, like `form.label` inside of a child component the
template object will be the original template object from where the form
builder was instantiated.

This means in cases where we pass a block to form builder helper the
`@output_buffer` is the parent component's output buffer. So when that
buffer is overridden via `.capture` it's setting the parent component's
`@output_buffer` to a new, temporary buffer but it doesn't change the
child component that is actually rendering's `@output_buffer`. This
means the block is executed in the context of the child components
actual `@output_buffer` and immediately writes to the buffer. Since
`capture` had nothing written to the temporary buffer, it returns the
value of the block call and then inserts that into the parent
component's `@output_buffer`.

This resolves the issue by monkey patching `ActionView::Base` to support
a new `child_component` attribute and overriding `output_buffer=` to set
its `child_component`s output_buffer to whatever `output_buffer` is set
to. With that in place, when `capture` is called and `#output_buffer=`
is called it will also update the child component's output buffer
correctly.