rename ActionView::Component to ViewComponent
This commit renames the bulk of the library to the ViewComponent
namespace, preserving backward compatibility with
ActionView::Component::Base and the associated base classes for now.
This change will allow users of the library to migrate to the next
major version ahead of the major v2 release.

ViewComponent::Base and its associated classes implement the
following changes:

1) Validations are not enabled by default.
2) Only the render(MyComponent.new) syntax is supported