Support Ruby 2.4 in CI
Update the `ruby_on_rails` action to re-add testing against Ruby 2.4.
To accomplish this, we need to run `gem update --system`, but we don't
want to do this on any version except `2.4.x` to save build time.

This commit uses the context and expression syntax for GitHub actions to
check if the Ruby versions starts with `2.4`, and update the system gems
in the event that we have a match.

We have to match like this because we specify `2.4.x` in our matrix but the
actual value of `matrix.ruby_version`, and the version that gets installed, is 2.4.9.
With this approach we have more flexibility if 2.4.x gets bumped (for some reason)
that we wouldn't get with a hard coded string.

Fixes #209