Replace the ActionView::Component migration description with some links
The information about the migration from ActionView::Component to
ViewComponent is not relevant for newcomers. To that end, it doesn't
seem necessary to continue taking up a large portion of the beginning of
the README with that information. Instead, we can link to the associated
issue and the old description from the README's history.