Avoid printing the missing capybara dependency warning by default
Instead, only print it if the `DEBUG` environment variable is set.

This removes unwanted noise for consumers of the gem who don't use
capybara and are fine with the limitation of not being able to use
capybara assertions. But if someone wants to use the capybara assertions
and isn't able to because they don't have a capybara dependency, they'll
be able to see the warning if they run their tests with `DEBUG=1`.