Add basic component and partial benchmarks
This adds a basic benchmark script using `benchmark/ips` to compare the
iterations per-second of component renders to partial renders.

This attempts to make the comparison equal by using the `test`
environments Rails application in production mode in addition to going
through a controllers `view_context` to render the partial and the
component.