Introduce component template inheritance
This change introduces template inheritance to components. Previously
subclassed components would have to define, or re-define a template
file. With this change, component templates are compiled when they're
inherited so the subclass will also inherit the `call_*` methods.

This functionality already existed for `inline` templates defined via
`call`, but this change introduces a test to ensure we have coverage of
that intended behavior.