Allow slots to accept content as argument or block
This aligns the interface between components and slots.

```
# This now works:
component = SlotsComponent.new
component.slot(:title).with_content { "This is my title!" }
component.slot(:subtitle).with_content { "This is my subtitle!" }
component.slot(:footer).with_content { "This is the footer" }
```

There is one exception: slots do not support passing components to `with_content`.

```
# NOT ALLOWED:
component = SlotsComponent.new
component.slot(:title).with_content(MyComponent.new)
```

This is because a slot does not know how to render things; it is just a PORO. It would be nice to offer this at a later time.