Simplify implementation
This wraps `with_content` args in a block, reducing conditionals in the implementation. In theory, this will add some overhead but, because of how minimal it was, it was difficult to measure. According to my benchmark, around 50 nanoseconds.

That is so small, I’m suspicious of my benchmark but I’d guess it is low enough not to matter yet.

```ruby
Benchmark.ips do |x|
  something = "doesn't matter what"
  value = "my value"
  block = -> (_something) { value }

  x.report("return value from block") do
    block.call(something)
  end

  x.report("return value") do
    value
  end

  x.compare!
end
```