Add support for sidecar translations
Expects a component to have a file with the same basename and
the .yml extension, holding translations scoped to the component's path.

A component YML should look like this:

  en:
    hello: "Hello World!"

And inside the component will be possible to call

  t(".hello") # => "Hello World!"

Internally the scope will be expanded using the template's path.
As an example, the translation above from within `app/components/example/component.html.erb`
will result in an internal I18n key like this:

  en:
    example:
      component:
        hello: "Hello World!"