Lazily evaluate content instead of eagerly
This changes `content` to load only when a component's `render?`
predicate method returns true. The `render?` method can still access
`content`, which will cause it to be evaluated.

This shouldn't be a breaking change for components themselves, but
attempting to access a slot in `before_render` or `render?` may no
longer work.