Decouple collection parameters from compilation
This change removes the logic that removed and recreated the
`collection_parameter`, `collection_counter_parameter`, and
`counter_argument_present?` methods on each compile from the compiler.

I don't know the history behind this logic living in the compiler, but
it looks like we may be able to define these as class methods on
`ViewComponent::Base` directly.

Thanks to @mejackreed for reporting the issue and helping write a
failing test.

closes https://github.com/github/view_component/issues/642