Fix handling of jbuilder templates
With jbuilder loaded the list of template handler extensions includes
both builder and jbuilder. The Compiler uses that list to build a glob
pattern which then contains '.*{builder,jbuilder}'. Since '.jbuilder'
matches both '*builder' and '*jbuilder' it is returned twice by Dir.[].
The Compiler will raise an error because it sees duplicate templates.
We prevent this issue by only returning a list of unique paths
in #matching_views_in_source_location.

Note that this issue is not specific to jbuilder. Any additional
template handler extensions which are a substring of an already existing
one will show the same behavior.