Ensure Base#request is available to ActionView asset helpers
ActionView::Helpers::AssetUrlHelper#compute_asset_host will try to
access the request with `respond_to?(:request)`, but since #request
was defined as private and the protocol option is forced to be
`:request` inside AV #asset_url it ended up trying to ask for a
protocol on nil.

This only happens if the asset_host is set to a hostname without a
defined protocol.

See:
- https://github.com/rails/rails/blob/v6.1.3.1/actionview/lib/action_view/helpers/asset_url_helper.rb#L277
- https://github.com/rails/rails/blob/v6.1.3.1/actionview/lib/action_view/helpers/asset_url_helper.rb#L231