Ensure session is set on request
Rails 7.0 deprecated writing to a session when the session is disabled
and defaults a session to a disabled session. https://github.com/rails/rails/pull/42231

In the past, ActionDispatch::TestRequest.create would set the
TestSession explicitly but this is not the case anymore and the default
session is used. With the PR mentioned above, this raises an exception
when testsing view components that implement sessions.