use Bundler v2 (#960)
I think it's time for us to move to Bundler v2. Having both
v1 and v2 in our CI workflow has been a source of headaches
and friction for some time. Seeing it trip up @boardfish
in https://github.com/github/view_component/pull/958 was
the last straw for me.

There are also now security concerns around Bundler v1
that to my knowledge are unresolved.

Removing Bundler v1 means that we can't easily test
Ruby 2.4 in CI. I'm OK with this, as we've never had a 2.4
compatibility issue caught by running against it.