Update tests for trailing newlines (#958)
Update tests with trailing newline solution

The tests introduced in #920 were made to pass by removing the trailing newline. These were supposed to check that trailing newlines in template files would not become part of the output. Now that the source issue has reached a conclusion thanks to an update to ActionView::Template::Handlers::ERB, the implementation for the test can be updated to reflect this. It also makes clear that we're dependent on that class for this change.