504 previews with sample source code (#667)
* Add method preview_example_source

Returns the method body from the preview file located in the host location under test/components/previews. This is being used if the preview file does not use render_template.

* Add layout for rendering source preview

* Add prism js and css assets

* Escape < and & and set language ruby and erb

* Add dependency method source

The gem adds the method #source which returns a method body as a string

* Make language name dynamic

Users can use other templating languages like haml. The template file extention can be used to retrieve the correct language. Fallback to ruby if there is not specific syntax highlighting available for the fiel extention.

* Update preview layout

-  use content_for for including code example within body
- remove stylesheet_link_tag. The stylesheet should be included explicitly via the host application.

* Add integration tests for example code preview

* Fix assets path

* Move assets to vendor folder

* Serve assets statically from gem bypassing assets pipeline or webpacker

* Shorten content_for key

* Prevent double yielding

yield was called twice when when content for view_component_preview_source is nil

* Partially revert preview layout related changes

* Rename preview tamplate helper to preview helper

* Introduce config option show_previews_source

Setting the option to true will add the preview source below non template previews.

* Add view helper preview_source

* Fix cops

* Rename preview_example_source to preview_source

* Return entire method source with signature

* Update changelog

* Update docs

* Fix markdown format

* Only return method body for preview

* Fix changelog

* Fix CHANGELOG

* Readd CHANGELOG symlink

* Add yard documentation

Co-authored-by: Hans Lemuet <Spone@users.noreply.github.com>
Co-authored-by: Hans Lemuet <hans@spone.net>