Add compatibility notes for using ActionText with ViewComponent (#990)
Summary

Using ActionText with ViewComponent is not supported out of the box. It results in an error.

This updates the compatibility documents with a workaround until it is supported.

This has been previously reported in these issues:

#445
#351

Both proposed solutions. The latter being the simplest in requiring the least configuration.

Thank you @coderberry and @andrewmcodes: #351 (comment)