Prevent nil from being returned from a Slot V2's to_s method (#1046)
When using `slim-rails` (and probably other temple-based languages), it's possible that
a block defined in a Slim view evaluates to `nil`. When that block gets processed in a
SlotV2's `to_s` method, that `nil` from the block can end up as the return value.

Returning `nil` from a call to `to_s` doesn't make sense and causes errors to be raised,
so this commit prevents that from happening and instead returns an empty string.