Extract ViewComponent::Core class
This extracts a new class, `ViewComponent::Core` which implements the
minimum amount of functionality to have a functioning component library.

The difference between `ViewComponent::Core` and `ViewComponent::Base`
is that Base includes extra functionality and compatibility with Rails,
such as slots, built-in route helpers, access to `request` and
`controller`, etc.

This new `Core` class is marked as private and should still be
considered experimental.