Add Orbit to the list of teams using ViewComponent
Hey! We’re very happy users of ViewComponent at [Orbit](https://orbit.love).
We’ve been using them for about a year and wrote about our [ViewComponent/Storybook setup](https://orbit.love/blog/building-a-component-library-in-rails-with-storybook) as well as a [primer on Storybook for Rails developers](https://orbit.love/blog/im-a-rails-developer-what-the-heck-is-storybook-and-should-i-care). We’re now in the process of building a design system with it, and the Primer repository as well as [this Rails Conf talk](https://youtu.be/QoetqsBCsbE) by Joel have been an invaluable help!