Separate slot setter and getter method
This introduces an explicit setter for slots in the form of
`with_#{slot_name}`. This makes it more clear when slots are being set
vs accessed but also allows chaining when slots have no arguments.

Before this change, the following would not work:

```ruby
component.header.with_content("my header!")
```

This fails due to `header` being passed no arguments and no block,
resulting in it acting as a getter, returning `nil` since the slot was
not defined. This is due to assuming that slots will always be passed
either a block or arguments and predates the `with_content` method.

Now, the above can be implemented using `with_*`:

```ruby
component.with_header.with_content("my header!")
```