Move changelog entry for #1162 to the right version
PR #1162 was merged in d24b26ae2697685e930d863839d7f8834b9ca4a3 which
was released in [v2.46.0] but its `CHANGELOG.me` entry had not been
moved above the `## v2.45.0` entry in
https://github.com/github/view_component/commit/a5f6463f36639e7385a9c905801e18e41b6b6538
so it was incorrectly listed as being included in v2.45.0.

Let's make sure it's listed under the correct version.

[v2.46.0]: https://github.com/github/view_component/releases/tag/v2.46.0