Allow component tests to define the current_user
Authentication gems like [clearance] and [devise] define a `current_user` helper
method that can be used in controllers and views to access the currently signed
in user.

When building components that depend on the currently signed in user and
delegating `:current_user` to the Rails helpers, it is useful to be able to
define `current_user` in tests.

The `with_current_user` helper method defines a current_user method on the
controller so that these components can be tested easily.

[clearance]: https://github.com/thoughtbot/clearance
[devise]: https://github.com/heartcombo/devise