Update CHANGELOG.md to provide a link to unreleased changes
We're going to automate the maintenance of the changelog because
maintaining it manually has been tedious and at times error-prone (for
both contributors and maintainers). Going forward, the changelog will be
updated when each release is created (via `script/release`), so it will
not include unreleased changes. But let's still provide a link to allow
folks to view the unreleased changes.