Add script to automate changelog updates during release
Up until now, updating the changelog has been a manual process for each
contribution. Since it was frequently forgotten, we added a GitHub
Actions job in 5f6d4235af9b9145bb9e68b6b9caa9622e950e46 to require each
PR to update the changelog. That aided reviewers (since they would no
longer have to check for a missing changelog entry), but contributors
were still left with the same burden of manually updating the changelog
and dealing with a few pain points:
- `CHANGELOG.md` is a frequent source of merge conflicts among the PRs
  that aren't merged within a short window
- Even when there isn't a conflict, merging `CHANGELOG.md` from `main`
  into a PR that's been open for a while can mistakenly place that PR's
  changelog entry under an already-released version (see
  07d434470c4e269c1afd91f78e0061683e6e7ce1)

Fortunately, I think we are in a good position to address these issues
by automating the changelog maintenance, especially because of the
conventions already in place in this repository:
- All PRs are required to have a changelog entry
- PRs are typically squash-merged into `main`

Because of those conventions, our changelog is very similar to the
repository's `git log`...so, I'm proposing we leverage the `git log` to
automatically generate our changelog at release-time. This will save
contributors from the burden of changelog management and provide some
additional benefits:
- Maintainers who want to alter a PR's changelog entry before merging
  can do so more easily: Previously, they would have to make edits to
  the code itself (which they can only do if the contributor has granted
  maintainers access to edit their branch). With a changelog generated
  from the git log, the maintainers can simply update the PR title (or,
  alternatively, the squashed commit message).
- Since squash-merging automatically appends the PR number to the
  generated commit's subject line, the generated changelog will
  automatically include references to each PR. (Previously, changelog
  entries did not include references to the corresponding PR.)