Support returning arrays from translations
When i18n finds a YAML array as a value, it returns the array as is.
ActionView allows this, and the #translate helper in AV will correctly
handle "HTML-safe" keys when the value is an Array.

This lifts the code from ActionView to handle HTML-safe translations so
we can correctly escape both strings and arrays.