Remove unreachable code branch
ActionView depends on ActiveSupport's SafeBuffer and its related core
extensions, which add #html_safe to Object. The check in the code we
took from ActionView looks to have been left there from earlier
iterations, before there was an explicit check in the code for Array
support.

See the discussion in this PR for more context and links to the relevant
commits in ActionView:

    https://github.com/github/view_component/pull/1290