Separate slot setter and getter method (#1184)
* Separate slot setter and getter method

This introduces an explicit setter for slots in the form of
`with_#{slot_name}`. This makes it more clear when slots are being set
vs accessed but also allows chaining when slots have no arguments.

Before this change, the following would not work:

```ruby
component.header.with_content("my header!")
```

This fails due to `header` being passed no arguments and no block,
resulting in it acting as a getter, returning `nil` since the slot was
not defined. This is due to assuming that slots will always be passed
either a block or arguments and predates the `with_content` method.

Now, the above can be implemented using `with_*`:

```ruby
component.with_header.with_content("my header!")
```

* Fix linter, add changelog entry

* Update lib/view_component/slotable_v2.rb

Co-authored-by: Hans Lemuet <Spone@users.noreply.github.com>

* Update lib/view_component/slotable_v2.rb

* Update docs/CHANGELOG.md

* Update docs, support polymorphic slots, add tests

* Fix ruby2_keywords call

Co-authored-by: Hans Lemuet <Spone@users.noreply.github.com>
Co-authored-by: Joel Hawksley <joel@hawksley.org>