Consistently evaluate passed in block content
Currently slots aren't composable due to the lifecycle of components,
where we attempt to avoid evaluating the block passed to `render_in`

(aka `content` in the library). This creates an inconsistent experience
where delegation works only if `content` was already evaluated.

The following code looks correct, but the `c.title` code will never be
executed:

```ruby
class MyComponent
  delegate :title, to: :@sidebar

  def initialize
    @sidebar = SidebarComponent.new
  end

  def call
    content_tag :div do
      render @sidebar
    end
  end
end
```

```erb
<%= render MyComponent.new(:title) do |c| %>
  <%= c.title("My sidebar") %>
<% end %>
```

The block passed that includes the call to `c.title("My sidebar")` is
never evaluated. This is a side-effect of a performance optimization
where we attempt to lazily evaluate the block passed to `render_in`.
This optimization was introduced to avoid evaluating  the entire render
chain for components return `false` when `render?` is called since it's
unnecessary work.

If the component were slightly different and called `content` for any
reason the above code would work as expected:

```ruby
class MyComponent
  delegate :title, to: :@sidebar

  def initialize
    @sidebar = SidebarComponent.new
  end

  def call
    # this causes `content` to be evaluated, resulting in
    # <%= c.title(title: "My sidebar") %> being called. Now this
    # component will work as expected.
    modified_content = content.replace('foo', 'bar')

    content_tag :div do
      render @sidebar
    end
  end
end
```

This is an example of the optimization causing inconsistent behavior in
the library. The change above causes `content` to be called before
`#call`, so any setup/mutations/side-effects are now guaranteed to be
reflected in the component output.

Here's what the change in the component lifecycle looks like:

```
(#content can be called anywhere in the above chain, or never called)
```

Now the lifecycle looks like:

```
(#content isn't used, it's just evaluated to ensure side-effects are performed)
```

This introduces the behavior change behind a configurable value that can
be set via `view_component.config.use_consistent_rendering_lifecycle`.

To enable this before it becomes the default behavior:

`config.view_component.use_consistent_rendering_lifecycle = true`

This is largely copied from
https://github.com/github/view_component/pull/1407 but introduces it as
a opt-in behavior instead of changing the default.