Fix request.path inside with_request_url helper
`#with_request_url` is great. It was started in #1058, but later #1221
introduced a tiny regression in the parsing of segments.

When using `#with_request_url` including a query string, the `#path`
helper would now contain the query string segment, when it should not.
This led to `#fullpath` having the query string twice.

This commit fixes the parsing and amends the test cases to cover this
scenario.