Apply the new standardrb rules (#1576)
While working on [supporting splatted keyword
args](https://github.com/ViewComponent/view_component/pull/1560) I
discovered that the linter was failing me on changes I didn't make.

When I ran standardrb locally, it was quite content, so I was a bit
confused. Then I read [.github/workflows/lint.yml] and realized that
view_component runs a `bundle update` prior to running standard; which
bumps the standardrb version and the corresponding rubocop set.

This wasintroduced by @camertron in
[e7eec9869fc13e401401c687573a98f1b0ed57a8](https://github.com/ViewComponent/view_component/commit/e7eec9869fc13e401401c687573a98f1b0ed57a8)
7+ months ago to fix an error in CI.

It's unclear to me why `bundle install` was not working, so I decided
that rather than swapping that back I would trust @camertron's instincts
and roll-forward with the update.

I am not including a change to the Changelog because this seems internal
and unnecessary to communicate to package consumers.