Prevent polymorphic slots causing content evaluation (#1575)
* Prevent polymorphic slots causing content evaluation

Currently there's a bug in polymorphic slots where setting a slot can
cause `content` to be evaluated early. This requires code to be order
dependent, setting content and other values before setting a polymorphic
slot. This primarily occurs in code that makes heavy use of component
API's within Ruby, not within typical ERB usage.

This updates the polymorphic slot code to avoid the `get_slot` method so
that content won't be evaluated early.

* Update via standardrb

* Add changelog entry

* Fix bad rebase

Co-authored-by: Joel Hawksley <joel@hawksley.org>