Don't count un-named splatted keyword args
In Ruby 3.1, it appears that the `ViewComponent::Base#initialize` method
has an implicit key-rest argument, named `:**`.

I thought it may be because we define it with a `def initialize(*); end;`
which could be implicitly adding in the rest of the splattable
parameter? However, playing around with that syntax doesn't seem to be
the cause:

```ruby
class A; def initialize(*); end; end
class B; end
A.instance_method(:initialize).parameters
B.instance_method(:initialize).parameters
```