`.with_collection` supports Components with splatted KeywordArgs (#1560)
* Initial support for Components with KeywordArgs

In a few projects, I wind up using keyword args on a Base class so that
I don't have to define an `initialize` on every child class. For
example:

```ruby
class ApplicationComponent < ViewComponent::Base
  def initialize(**kwargs)
    super
    kwargs.each do |k, v|
      send(:"#{k}=", v)
    end
  end
end

class ProductComponent < ApplicationComponent
  attr_accessor :product
end

class OrderComponent < ApplicationComponent
  attr_accessor :order
end

```

However, ViewComponent doesn't recognize that a keyword args is an
implicit `def initialize(product:)` or `def initialize(order:)`; which
causes it to raise an `ArgumentError` for an unaccepted parameter.

This causes Components with Keyword Args in their initialize to pass the
initialze parameter name check.

Open Questions:

- [ ] Are there side-effects of this I'm not aware of?
- [ ] Does this create subtle design pressure that encourages cruft?

Thoughts?

* Render nothing just to get past the explode boom

* Lint lint lint

* Standardrb vs Ruby 2.5: FIGHT!

* Add Changelog per CI

* Update docs/CHANGELOG.md

Co-authored-by: Joel Hawksley <joelhawksley@github.com>

* Assert rendering against the collection works

* The Kwarg is the Component name 🙈

I appear to have forgotten that the with_collection leverages the
components name to figure out the name of the argument it uses for the
collection; rather than the name of the items in the collection (which
is how it works in Rails).

* Clarify that this is for Splatted Keyword Args

Co-authored-by: jcoyne@users.noreply.github.com

* Keyword argument detection is more resilient

* Don't count un-named splatted keyword args

In Ruby 3.1, it appears that the `ViewComponent::Base#initialize` method
has an implicit key-rest argument, named `:**`.

I thought it may be because we define it with a `def initialize(*); end;`
which could be implicitly adding in the rest of the splattable
parameter? However, playing around with that syntax doesn't seem to be
the cause:

```ruby
class A; def initialize(*); end; end
class B; end
A.instance_method(:initialize).parameters
B.instance_method(:initialize).parameters
```

* Make it actually private

Co-authored-by: Joel Hawksley <joelhawksley@github.com>