Add TestHelpers#vc_test_controller
In https://github.com/ViewComponent/view_component/pull/1661, I
renamed several private TestHelper methods to indicate that
they should not be used by consumers.

However, we depended on the `controller` method in the GitHub.com
test suite. I've chosen to reintroduce this method with new name,
as we originally made it private due to it conflicting with folks
setting a `controller` method name on tests.