Add TestHelpers#vc_test_request (#1678)
In #1661, I renamed several private TestHelper methods to indicate that they should not be used by consumers.

However, we depended on the request method in the GitHub.com test suite. I've chosen to reintroduce this method with a new name, as we originally made it private due to it conflicting with folks setting a request method name in tests.